requirements.txt
django-unfold>= 0.30.0
django-svelte-jsoneditor>= 0.4.2
django-import-export>= 4.0.3
django-admin-confirm>= 1.0.0
